+++ packages/w3wallets/src/metamask/metamask.ts
@@ -1,4 +1,4 @@
-import { type Page } from "@playwright/test";
+import { expect, type Page } from "@playwright/test";
 import { Wallet } from "../wallet";
 import type { NetworkSettings } from "./types";
 
@@ -32,33 +32,99 @@ export class Metamask extends Wallet {
 
     await this.page.getByTestId("pin-extension-next").click();
     await this.page.getByTestId("pin-extension-done").click();
+
+    // Sometimes MM uses dialogues – close them
+    await this.page.waitForTimeout(1000);
+    await this.clickTopRightCornerToCloseAllTheMarketingBullshit();
   }
 
-  async connectToNetwork(settings: NetworkSettings, switchNetwork = true) {
-    await this.page.locator(".mm-picker-network").click();
+  async switchAccount(nameOrAddress: string, network: "ETH" | "SOL" = "ETH") {
+    if (nameOrAddress.startsWith("0x"))
+      nameOrAddress = nameOrAddress.slice(0, 7);
+    await this.page.getByTestId("account-menu-icon").click();
     await this.page
-      .getByRole("button", { name: "Add a custom network" })
+      .locator(".multichain-account-menu-popover__list--menu-item")
+      .filter({ hasText: nameOrAddress })
+      .filter({
+        has: this.page
+          .getByTestId("second-currency-display")
+          .filter({ hasText: network }),
+      })
+      .click();
+  }
+
+  async importAccount(privateKey: string) {
+    await this.page.getByTestId("account-menu-icon").click();
+    await this.page
+      .getByTestId("multichain-account-menu-popover-action-button")
       .click();
     await this.page
-      .getByTestId("network-form-network-name")
-      .fill(settings.name);
-    await this.page
-      .getByTestId("network-form-chain-id")
-      .fill(settings.chainId.toString());
-    await this.page
-      .getByTestId("network-form-ticker-input")
-      .fill(settings.currencySymbol);
+      .getByTestId("multichain-account-menu-popover-add-imported-account")
+      .click();
+    await this.page.locator("#private-key-box").fill(privateKey);
+    await this.page.getByTestId("import-account-confirm-button").click();
+  }
 
-    await this.page.getByTestId("test-add-rpc-drop-down").click();
-    await this.page.getByRole("button", { name: "Add RPC URL" }).click();
-    await this.page.getByTestId("rpc-url-input-test").fill(settings.rpc);
-    await this.page.getByRole("button", { name: "Add URL" }).click();
-    await this.page.getByRole("button", { name: "Save" }).click();
-
-    if (switchNetwork) {
-      await this.page.locator(".mm-picker-network").click();
-      await this.page.getByTestId(settings.name).click();
+  async addAccount(accountName?: string) {
+    await this.page.getByTestId("account-menu-icon").click();
+    await this.page
+      .getByTestId("multichain-account-menu-popover-action-button")
+      .click();
+    await this.page
+      .getByTestId("multichain-account-menu-popover-add-account")
+      .click();
+    if (accountName) {
+      await this.page.locator("#account-name").fill(accountName);
     }
+    await this.page.getByTestId("submit-add-account-with-name").click();
+  }
+
+  async getAccountName() {
+    const accountSelect = this.page.getByTestId("account-menu-icon");
+    await expect(accountSelect).toBeVisible();
+    const text = await accountSelect.textContent();
+    if (!text) throw Error("Cannot get account name");
+    return text;
+  }
+
+  async connectToNetwork(networkName: string): Promise<void>;
+  async connectToNetwork(settings: NetworkSettings): Promise<void>;
+  async connectToNetwork(settingsOrName: NetworkSettings | string) {
+    if (typeof settingsOrName !== "string") {
+      await this.page.locator(".mm-picker-network").click();
+      await this.page
+        .getByRole("button", { name: "Add a custom network" })
+        .click();
+      await this.page
+        .getByTestId("network-form-network-name")
+        .fill(settingsOrName.name);
+      await this.page
+        .getByTestId("network-form-chain-id")
+        .fill(settingsOrName.chainId.toString());
+      await this.page
+        .getByTestId("network-form-ticker-input")
+        .fill(settingsOrName.currencySymbol);
+
+      await this.page.getByTestId("test-add-rpc-drop-down").click();
+      await this.page.getByRole("button", { name: "Add RPC URL" }).click();
+      await this.page
+        .getByTestId("rpc-url-input-test")
+        .fill(settingsOrName.rpc);
+      await this.page.getByRole("button", { name: "Add URL" }).click();
+      await this.page.getByRole("button", { name: "Save" }).click();
+    }
+
+    await this.page.locator(".mm-picker-network").click();
+    await this.page
+      .locator("text=Show test networks >> xpath=following-sibling::label")
+      .click();
+    await this.page
+      .getByTestId(
+        typeof settingsOrName === "string"
+          ? settingsOrName
+          : settingsOrName.name,
+      )
+      .click();
   }
 
   // async approve() {
@@ -93,5 +159,9 @@ export class Metamask extends Wallet {
     await p.goto(`chrome-extension://${this.extensionId}/notification.html`);
     await action(p);
     await p.close();
+  }
+
+  private async clickTopRightCornerToCloseAllTheMarketingBullshit() {
+    await this.page.mouse.click(1000, 10);
   }
 }